Add AvalableDoctors component, MapComponent refactor
