+++ frontend/src/components/SearchClinicsByAddressForm.jsx
@@ -1,0 +1,37 @@
+import React, { useState, useEffect } from 'react';
+import { GoogleMap, useJsApiLoader, Marker } from '@react-google-maps/api';
+import axios from 'axios';
+import ClinicsList from './ClinicsList/ClinicsList';
+
+
+const SearchClinicsByAddressForm = ({setCoordinates, setSearchTermMarker}) => {
+  const [searchTerm, setSearchTerm] = useState('');
+
+  const handleSearchByAddress = (e) => {
+    e.preventDefault();
+  
+    axios.get(`clinics/api/geocode?address=${searchTerm}`)
+      .then(response => {
+        const { latitude, longitude } = response.data;
+        setCoordinates({ lat: latitude, lng: longitude });
+        setSearchTermMarker(true);
+      })
+      .catch(error => {
+        console.error('Error geocoding address:', error);
+      });
+  };
+  
+  return (
+    <form onSubmit={handleSearchByAddress}>
+      <input
+        type="text"
+        value={searchTerm}
+        onChange={(e) => setSearchTerm(e.target.value)}
+        placeholder="Enter your zip code" 
+      />
+      <button>Search</button>
+    </form>
+  );
+};
+
+export default SearchClinicsByAddressForm;
